pyproject.toml
[tool.poetry]
name = "recipe-recommender"
version = "0.1.0"
description = ""
authors = ["Bucky Roberts <buckyroberts@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.28.0"
llama-index = "^0.8.29.post1"
promptlayer = "^0.2.9"
pinecone-client = "^2.2.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
